dump

+++ srcs/complexes/ft_get_complexe_magnitude.c
@@ -6,7 +6,7 @@
 /*   By: sclolus <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/07/03 07:54:39 by sclolus           #+#    #+#             */
-/*   Updated: 2017/07/03 07:56:39 by sclolus          ###   ########.fr       */
+/*   Updated: 2017/07/15 00:51:24 by sclolus          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,5 @@
 
 double	ft_get_complexe_magnitude(t_complexe *z)
 {
-	return (sqrt(pow(z->real_part, 2) + pow(z->imaginary_part, 2)));
+	return ((double)sqrt(pow((double)z->real_part, 2) + (double)pow((double)z->imaginary_part, 2)));
 }
